+++ home.html
@@ -104,9 +104,11 @@
             </ul>
         </footer>
 
-        <script src="js/posts.js"></script>
-
-        <script src="js/user/signout.js"></script>
+        <!-- <script src="js/posts.js"></script> -->
+        <script src="js/modules/posts/allPosts.mjs" type="module"></script>
+        <!-- <script src="js/home.js" type="module"></script> -->
+        <script src="js/createPosts.js"></script>
+        <script src="js/modules/user/signout.js"></script>
 
     </body>
